requirements.txt
matplotlib==2.2.2
numpy==1.17.2
opencv-contrib-python==4.0.0.21
opencv-python==4.1.2.30
opencv-python-headless==4.1.2.30
pandas==0.22.0
Pillow==6.1.0
pyparsing==2.4.0
jsonschema==3.0.1
scikit-image==0.15.0
scikit-learn==0.19.1
scipy==1.3.1
sk-video==1.1.10
tb-nightly==2.2.0a20191211
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
torch==1.3.1
torchvision==0.4.2
tqdm==4.41.0